requirements.txt
markdown2==2.4.10
Pygments==2.16.1
python-slugify==8.0.1
text-unidecode==1.3
